+++ src/main/java/mx/ecommerce/model/Producto.java
@@ -26,14 +26,14 @@ public class Producto implements java.io.Serializable {
 	private static final long serialVersionUID = 1L;
 	private Integer id;
 	private String nombre;
-	private Float cantidad;
+	private Integer cantidad;
 	private Float precio;
 	private Set<CategoriaProducto> categorias = new HashSet<CategoriaProducto>(0);
 
 	public Producto() {
 	}
 
-	public Producto(String nombre, Float cantidad, Float precio) {
+	public Producto(String nombre, Integer cantidad, Float precio) {
 		this.nombre = nombre;
 		this.cantidad = cantidad;
 		this.precio = precio;
@@ -70,11 +70,11 @@ public class Producto implements java.io.Serializable {
 	}
 	
 	@Column(name = "cantidad", unique = false, nullable = false)
-	public Float getCantidad() {
+	public Integer getCantidad() {
 		return cantidad;
 	}
 
-	public void setCantidad(Float cantidad) {
+	public void setCantidad(Integer cantidad) {
 		this.cantidad = cantidad;
 	}
